Delete source code of STL-Vector realization. Move solution to Source.cpp
